PatientRoom service update

+++ SerenityHospital.Business/Services/Implements/PatientRoomService.cs
@@ -1,9 +1,11 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Identity;
 using SerenityHospital.Business.Constants;
 using SerenityHospital.Business.Dtos.PatientRoomDtos;
 using SerenityHospital.Business.Exceptions.Common;
 using SerenityHospital.Business.Exceptions.Images;
 using SerenityHospital.Business.Exceptions.PatientRooms;
+using SerenityHospital.Business.Exceptions.Patients;
 using SerenityHospital.Business.Extensions;
 using SerenityHospital.Business.ExternalServices.Interfaces;
 using SerenityHospital.Business.Services.Interfaces;
@@ -18,13 +20,15 @@ public class PatientRoomService : IPatientRoomService
     readonly IDepartmentRepository _depRepo;
     readonly IMapper _mapper;
     readonly IFileService _fileService;
+    readonly UserManager<Patient> _patientUserManager;
 
-    public PatientRoomService(IPatientRoomRepository repo, IMapper mapper, IDepartmentRepository depRepo, IFileService fileService)
+    public PatientRoomService(IPatientRoomRepository repo, IMapper mapper, IDepartmentRepository depRepo, IFileService fileService, UserManager<Patient> patientUserManager)
     {
         _repo = repo;
         _mapper = mapper;
         _depRepo = depRepo;
         _fileService = fileService;
+        _patientUserManager = patientUserManager;
     }
 
     public async Task CreateAsync(PatientRoomCreateDto dto)
@@ -47,8 +51,10 @@ public class PatientRoomService : IPatientRoomService
     public async Task DeleteAsync(int id)
     {
         if (id <= 0) throw new NegativeIdException<PatientRoom>();
-        var entity = await _repo.GetByIdAsync(id);
+        var entity = await _repo.GetByIdAsync(id,"Patients");
         if (entity is null) throw new NotFoundException<PatientRoom>();
+
+        if (entity.Patients.Count() > 0) throw new PatientRoomNotEmptyException();
 
         _repo.Delete(entity);
         _fileService.Delete(entity.ImageUrl);
@@ -59,11 +65,11 @@ public class PatientRoomService : IPatientRoomService
     {
         if(takeAll)
         {
-            return _mapper.Map<IEnumerable<PatientRoomListItemDto>>(_repo.GetAll());
+            return _mapper.Map<IEnumerable<PatientRoomListItemDto>>(_repo.GetAll("Patients"));
         }
         else
         {
-            return _mapper.Map<IEnumerable<PatientRoomListItemDto>>(_repo.FindAll(pr => pr.IsDeleted == false));
+            return _mapper.Map<IEnumerable<PatientRoomListItemDto>>(_repo.FindAll(pr => pr.IsDeleted == false, "Patients"));
         }
     }
 
@@ -75,12 +81,12 @@ public class PatientRoomService : IPatientRoomService
 
         if(takeAll)
         {
-            entity =await _repo.GetByIdAsync(id);
+            entity =await _repo.GetByIdAsync(id,"Patients");
             if (entity is null) throw new NotFoundException<PatientRoom>();
         }
         else
         {
-            entity = await _repo.GetSingleAsync(pr => pr.IsDeleted == false && pr.Id==id);
+            entity = await _repo.GetSingleAsync(pr => pr.IsDeleted == false && pr.Id==id, "Patients");
             if (entity is null) throw new NotFoundException<PatientRoom>();
         }
 
@@ -100,8 +106,10 @@ public class PatientRoomService : IPatientRoomService
     public async Task SoftDeleteAsync(int id)
     {
         if (id <= 0) throw new NegativeIdException<PatientRoom>();
-        var entity = await _repo.GetByIdAsync(id);
+        var entity = await _repo.GetByIdAsync(id,"Patients");
         if (entity is null) throw new NotFoundException<PatientRoom>();
+
+        if (entity.Patients.Count() > 0) throw new PatientRoomNotEmptyException();
 
         _repo.SoftDelete(entity);
         await _repo.SaveAsync();
@@ -110,7 +118,7 @@ public class PatientRoomService : IPatientRoomService
     public async Task UpdateAsync(int id, PatientRoomUpdateDto dto)
     {
         if (id <= 0) throw new NegativeIdException<PatientRoom>();
-        var entity = await _repo.GetByIdAsync(id);
+        var entity = await _repo.GetByIdAsync(id,"Patients");
         if (entity is null) throw new NotFoundException<PatientRoom>();
 
         if (await _repo.IsExistAsync(pr => pr.Number == dto.Number && pr.Id != id)) throw new ThisRoomNumberIsAlreadyExistException();
@@ -127,6 +135,25 @@ public class PatientRoomService : IPatientRoomService
         var department =await _depRepo.GetByIdAsync(dto.DepartmentId);
         if (department is null) throw new NotFoundException<Department>();
 
+        entity.Patients?.Clear();
+
+        if(dto.Patientids !=null)
+        {
+            foreach (var patientId in dto.Patientids)
+            {
+                var patient = await _patientUserManager.FindByIdAsync(patientId);
+                if (patient is null) throw new NotFoundException<Patient>();
+
+                var patientInOtherPatientRoom = await _repo.IsExistAsync(p => p.Id != id && p.Patients.Any(p => p.Id == patient.Id));
+                if (patientInOtherPatientRoom) throw new PatientInOtherPatientRoomException();
+
+
+                if (dto.Patientids.Count() > dto.Capacity) throw new PatientRoomCapacityIsFullException();
+
+                entity.Patients?.Add(patient);
+            }
+
+        }
         entity.DepartmentId = dto.DepartmentId;
 
         _mapper.Map(dto, entity);
